Language.SOAS.Impl.Generated

Documentation

>>> :set -XOverloadedStrings
>>> :set -XDataKinds
>>> import qualified Control.Monad.Foil as Foil
>>> import qualified Language.SOAS.Syntax.Abs as Raw
>>> import Control.Monad.Free.Foil
>>> import Data.String (fromString)

Generated code

type ScopedOpArgTyping' a (n :: S) = ScopedOpArgTyping'Sig a (ScopedAST (TypeBinders' a) (Type'Sig a) n) (AST (TypeBinders' a) (Type'Sig a) n)

Generated with mkFreeFoil. A scope-safe version of ScopedOpArgTyping'.

data ScopedOpArgTyping'Sig a scope term where

Generated with mkFreeFoil. A signature based on ScopedOpArgTyping'.

Constructors

ScopedOpArgTypingSig :: forall a scope term. a -> OpArgTyping'Sig a scope term -> ScopedOpArgTyping'Sig a scope term
Corresponds to ScopedOpArgTyping.

type OpArgTyping' a (n :: S) = OpArgTyping'Sig a (ScopedAST (TypeBinders' a) (Type'Sig a) n) (AST (TypeBinders' a) (Type'Sig a) n)

Generated with mkFreeFoil. A scope-safe version of OpArgTyping'.

data OpArgTyping'Sig a scope term where

Generated with mkFreeFoil. A signature based on OpArgTyping'.

Constructors

OpArgTypingSig :: forall a term scope. a -> [term] -> term -> OpArgTyping'Sig a scope term
Corresponds to OpArgTyping.

type ScopedType' a = ScopedAST (TypeBinders' a) (Type'Sig a)

Generated with mkFreeFoil. A scoped (and scope-safe) version of Type'.

type Type' a = AST (TypeBinders' a) (Type'Sig a)

Generated with mkFreeFoil. A scope-safe version of Type'.

data Type'Sig a scope term where

Generated with mkFreeFoil. A signature based on Type'.

Constructors

TypeFunSig :: forall a term scope. a -> term -> term -> Type'Sig a scope term
Corresponds to TypeFun.

TypeProductSig :: forall a term scope. a -> term -> term -> Type'Sig a scope term
Corresponds to TypeProduct.

type OpArg' a (n :: S) = OpArg'Sig a (ScopedAST (Binders' a) (Term'Sig a) n) (AST (Binders' a) (Term'Sig a) n)

Generated with mkFreeFoil. A scope-safe version of OpArg'.

data OpArg'Sig a scope term where

Generated with mkFreeFoil. A signature based on OpArg'.

Constructors

OpArgSig :: forall a scope term. a -> scope -> OpArg'Sig a scope term
Corresponds to OpArg.

PlainOpArgSig :: forall a term scope. a -> term -> OpArg'Sig a scope term
Corresponds to PlainOpArg.

type ScopedTerm' a = ScopedAST (Binders' a) (Term'Sig a)

Generated with mkFreeFoil. A scoped (and scope-safe) version of Term'.

type Term' a = AST (Binders' a) (Term'Sig a)

Generated with mkFreeFoil. A scope-safe version of Term'.

data Term'Sig a scope term where

Generated with mkFreeFoil. A signature based on Term'.

Constructors

OpSig :: forall a scope term. a -> OpIdent -> [OpArg'Sig a scope term] -> Term'Sig a scope term
Corresponds to Op.

MetaVarSig :: forall a term scope. a -> MetaVarIdent -> [term] -> Term'Sig a scope term
Corresponds to MetaVar.

data TypeBinders' a (o :: S) (i :: S) where

Generated with mkFreeFoil. A binding type, scope-safe version of TypeBinders'.

Constructors

NoTypeBinders :: forall a (o :: S). a -> TypeBinders' a o o
Corresponds to NoTypeBinders.

SomeTypeBinders :: forall a (o :: S) (i1 :: S) (i :: S). a -> NameBinder o i1 -> TypeBinders' a i1 i -> TypeBinders' a o i
Corresponds to SomeTypeBinders.

data Binders' a (o :: S) (i :: S) where

Generated with mkFreeFoil. A binding type, scope-safe version of Binders'.

Constructors

NoBinders :: forall a (o :: S). a -> Binders' a o o
Corresponds to NoBinders.

SomeBinders :: forall a (o :: S) (i1 :: S) (i :: S). a -> NameBinder o i1 -> Binders' a i1 i -> Binders' a o i
Corresponds to SomeBinders.

data TermTyping' a (o :: S) where

Generated with mkFreeFoil. A scope-safe version of TermTyping'.

Constructors

TermTyping :: forall a (o :: S) (i :: S). a -> TypeBinders' a o i -> Context' a -> Term' a i -> Type' a i -> TermTyping' a o
Corresponds to TermTyping.

data VarTyping' a (o :: S) where

Generated with mkFreeFoil. A scope-safe version of VarTyping'.

Constructors

VarTyping :: forall a (o :: S). a -> Name o -> Type' a o -> VarTyping' a o
Corresponds to VarTyping.

data Constraint' a (o :: S) where

Generated with mkFreeFoil. A scope-safe version of Constraint'.

Constructors

ConstraintEq :: forall a (o :: S) (i :: S). a -> Binders' a o i -> Term' a i -> Term' a i -> Constraint' a o
Corresponds to ConstraintEq.

data OpTyping' a (o :: S) where

Generated with mkFreeFoil. A scope-safe version of OpTyping'.

Constructors

OpTyping :: forall a (o :: S) (i :: S). a -> OpIdent -> TypeBinders' a o i -> [ScopedOpArgTyping' a i] -> Type' a i -> OpTyping' a o
Corresponds to OpTyping.

data MetaVarTyping' a (o :: S) where

Generated with mkFreeFoil. A scope-safe version of MetaVarTyping'.

Constructors

MetaVarTyping :: forall a (o :: S). a -> MetaVarIdent -> [Type' a o] -> Type' a o -> MetaVarTyping' a o
Corresponds to MetaVarTyping.

data Subst' a (o :: S) where

Generated with mkFreeFoil. A scope-safe version of Subst'.

Constructors

Subst :: forall a (o :: S) (i :: S). a -> MetaVarIdent -> Binders' a o i -> Term' a i -> Subst' a o
Corresponds to Subst.

pattern ScopedOpArgTyping :: forall a (o :: S). a -> OpArgTyping' a o -> ScopedOpArgTyping' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type ScopedOpArgTyping.

pattern OpArgTyping :: forall a (o :: S). a -> [Type' a o] -> Type' a o -> OpArgTyping' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type OpArgTyping.

pattern TypeProduct :: a -> Type' a o -> Type' a o -> Type' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type TypeProduct.

pattern TypeFun :: a -> Type' a o -> Type' a o -> Type' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type TypeFun.

pattern PlainOpArg :: forall a (o :: S). a -> Term' a o -> OpArg' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type PlainOpArg.

pattern OpArg :: forall a (o :: S) (i :: S). () => a -> Binders' a o i -> Term' a i -> OpArg' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type OpArg.

pattern MetaVar :: a -> MetaVarIdent -> [Term' a o] -> Term' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type MetaVar.

pattern Op :: a -> OpIdent -> [OpArg' a o] -> Term' a o

Generated with mkFreeFoil. Pattern synonym for an AST node of type Op.

fromTerm'Sig :: Term'Sig a (Binders' a, ScopedTerm' a) (Term' a) -> Term' a

Generated with mkFreeFoil. A helper used to convert from scope-safe to raw representation.

fromOpArg'Sig :: OpArg'Sig a (Binders' a, ScopedTerm' a) (Term' a) -> OpArg' a

Generated with mkFreeFoil. A helper used to convert from scope-safe to raw representation.

fromBinders' :: forall a (o :: S) (i :: S). Binders' a o i -> Binders' a

Generated with mkFreeFoil. Convert a scope-safe to a raw binding.

fromOpArg' :: forall a (o :: S). OpArg' a o -> OpArg' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromTerm' :: forall a (o :: S). Term' a o -> Term' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromType'Sig :: Type'Sig a (TypeBinders' a, ScopedType' a) (Type' a) -> Type' a

Generated with mkFreeFoil. A helper used to convert from scope-safe to raw representation.

fromOpArgTyping'Sig :: OpArgTyping'Sig a (TypeBinders' a, ScopedType' a) (Type' a) -> OpArgTyping' a

Generated with mkFreeFoil. A helper used to convert from scope-safe to raw representation.

fromScopedOpArgTyping'Sig :: ScopedOpArgTyping'Sig a (TypeBinders' a, ScopedType' a) (Type' a) -> ScopedOpArgTyping' a

Generated with mkFreeFoil. A helper used to convert from scope-safe to raw representation.

fromTypeBinders' :: forall a (o :: S) (i :: S). TypeBinders' a o i -> TypeBinders' a

Generated with mkFreeFoil. Convert a scope-safe to a raw binding.

fromOpArgTyping' :: forall a (o :: S). OpArgTyping' a o -> OpArgTyping' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromScopedOpArgTyping' :: forall a (i :: S). ScopedOpArgTyping' a i -> ScopedOpArgTyping' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromType' :: forall a (o :: S). Type' a o -> Type' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromSubst' :: forall a (o :: S). Subst' a o -> Subst' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromMetaVarTyping' :: forall a (o :: S). MetaVarTyping' a o -> MetaVarTyping' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromOpTyping' :: forall a (o :: S). OpTyping' a o -> OpTyping' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromConstraint' :: forall a (o :: S). Constraint' a o -> Constraint' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromVarTyping' :: forall a (o :: S). VarTyping' a o -> VarTyping' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

fromTermTyping' :: forall a (o :: S). TermTyping' a o -> TermTyping' a

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toTerm'Sig :: Term' a -> Either VarIdent (Term'Sig a (Binders' a, ScopedTerm' a) (Term' a))

Generated with mkFreeFoil. A helper used to convert from raw to scope-safe representation.

toOpArg'Sig :: OpArg' a -> OpArg'Sig a (Binders' a, ScopedTerm' a) (Term' a)

Generated with mkFreeFoil. A helper used to convert from raw to scope-safe representation.

toBinders' :: forall (o :: S) a r. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> Binders' a -> (forall (i :: S). DExt o i => Binders' a o i -> Map VarIdent (Name i) -> r) -> r

Generated with mkFreeFoil. Convert from raw to scope-safe binding (CPS-style).

toOpArg' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> OpArg' a -> OpArg' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toTerm' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> Term' a -> Term' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toType'Sig :: Type' a -> Either VarIdent (Type'Sig a (TypeBinders' a, ScopedType' a) (Type' a))

Generated with mkFreeFoil. A helper used to convert from raw to scope-safe representation.

toOpArgTyping'Sig :: OpArgTyping' a -> OpArgTyping'Sig a (TypeBinders' a, ScopedType' a) (Type' a)

Generated with mkFreeFoil. A helper used to convert from raw to scope-safe representation.

toScopedOpArgTyping'Sig :: ScopedOpArgTyping' a -> ScopedOpArgTyping'Sig a (TypeBinders' a, ScopedType' a) (Type' a)

Generated with mkFreeFoil. A helper used to convert from raw to scope-safe representation.

toTypeBinders' :: forall (o :: S) a r. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> TypeBinders' a -> (forall (i :: S). DExt o i => TypeBinders' a o i -> Map VarIdent (Name i) -> r) -> r

Generated with mkFreeFoil. Convert from raw to scope-safe binding (CPS-style).

toOpArgTyping' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> OpArgTyping' a -> OpArgTyping' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toScopedOpArgTyping' :: forall (i :: S) a. (Distinct i, Ord VarIdent) => Scope i -> Map VarIdent (Name i) -> ScopedOpArgTyping' a -> ScopedOpArgTyping' a i

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toType' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> Type' a -> Type' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toSubst' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> Subst' a -> Subst' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toMetaVarTyping' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> MetaVarTyping' a -> MetaVarTyping' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toOpTyping' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> OpTyping' a -> OpTyping' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toConstraint' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> Constraint' a -> Constraint' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toVarTyping' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> VarTyping' a -> VarTyping' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

toTermTyping' :: forall (o :: S) a. (Distinct o, Ord VarIdent) => Scope o -> Map VarIdent (Name o) -> TermTyping' a -> TermTyping' a o

Generated with mkFreeFoil. Convert from scope-safe to raw representation.

User-defined helpers

unsafeParse :: ([Token] -> Either String a) -> String -> a

Orphan instances

ZipMatchK BNFC'Position
Ignore BNFC'Position when matching terms.

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (BNFC'Position :@@: as) -> (BNFC'Position :@@: bs) -> Maybe (BNFC'Position :@@: cs)

ZipMatchK MetaVarIdent
Match MetaVarIdent via Eq.

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (MetaVarIdent :@@: as) -> (MetaVarIdent :@@: bs) -> Maybe (MetaVarIdent :@@: cs)

ZipMatchK OpIdent
Match OpIdent via Eq.

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (OpIdent :@@: as) -> (OpIdent :@@: bs) -> Maybe (OpIdent :@@: cs)